// src/models/Ticket.ts
export interface Ticket {
    reimbId: number;
    reimbAmount: number;
    reimbSubmitted: string;
    reimbResolved: string;
    reimbDescription: string;
    reimbReceipt: string;
    reimbAuthor: number;
    reimbResolver: number;
    reimbStatusId: number;
    reimbTypeId: number;
    reimbStatus: string;
}